+++ packages/elements/src/components/Docs/index.tsx
@@ -1,8 +1,8 @@
 import { Dictionary, NodeType } from '@stoplight/types';
 import * as React from 'react';
 
-import { useLocationHash } from '../../hooks/useLocationHash';
 import { useParsedData } from '../../hooks/useParsedData';
+import { getLocationHash, scrollToAnchor } from '../../utils/url';
 import { Article } from './Article';
 import { HttpOperation } from './HttpOperation';
 import { HttpService } from './HttpService';
@@ -53,13 +53,7 @@ export const Docs = React.memo<IDocsProps>(({ nodeType, nodeData, className, hea
 export const ParsedDocs: React.FC<IParsedDocsProps> = ({ nodeType, nodeData, className, headless }) => {
   const Component = NodeTypeComponent[nodeType];
 
-  const locationHash = useLocationHash();
-
-  const scrollToAnchor = (hash: string) => {
-    if (hash && typeof document !== undefined) {
-      window.requestAnimationFrame(() => document.getElementById(hash)?.scrollIntoView(true));
-    }
-  };
+  const locationHash = getLocationHash();
 
   React.useEffect(() => {
     scrollToAnchor(locationHash);
